requirements.txt
fastapi
uvicorn
httpx
pydantic
SQLAlchemy
PyJWT
passlib[bcrypt]
email-validator
bcrypt
pydantic[email]
pytest
